requirements.txt
# Core Dependencies
boto3>=1.28.0           # AWS SDK
google-cloud-compute>=1.13.0  # GCP SDK
azure-mgmt-compute>=30.0.0    # Azure Compute SDK
azure-identity>=1.13.0        # Azure Authentication

# CLI Framework
typer[all]>=0.9.0      # Modern CLI framework
pydantic>=2.0.0        # Data validation

# Utilities
python-dotenv>=1.0.0   # Environment variables
rich>=13.0.0           # Terminal formatting
tabulate>=0.9.0        # Table formatting

# Development & Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0          # Code formatter
flake8>=6.0.0          # Linter
mypy>=1.0.0            # Type checker
